requirements.txt
pywhispercpp>=1.0.0
ffmpeg-python>=0.2.0
deep-translator>=1.8.0
tqdm>=4.60.0
streamlit>=1.30.0
coremltools>=7.0
ane_transformers
openai-whisper
